+++ src/core/utils/index-script.ts
@@ -5,7 +5,7 @@ export { randomId } from './id.js';
 export { escapeHtml, numDelim } from './html.js';
 export type { TypedArray } from './is-type.js';
 export { isTypedArray, isArray, isSet, isRegExp, isError } from './is-type.js';
-export { jsonStringifyAsJavaScript, jsonStringifyInfo } from './json.js';
+export { jsonStringifyAsJavaScript, jsonStringifyInfo, jsonSafeParse, jsonSafeStringify } from './json.js';
 export type * from './load-data.js';
 export * from './load-data.js';
 export type * from './logger.js';