+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## next
 
 - Bumped [jora](https://github.com/discoveryjs/jora) to 1.0.0-beta.15
+- Added `jsonSafeParse()` and `jsonSafeStringify()` utility functions, which behave like `JSON.parse()` and `JSON.stringify()` but return an error instead of throwing an exception
 - Added `isError()` utility function, serving as an alias for `Error.isError()` with a fallback if unavailable
 - Added support for displaying error values (where `isError()` returns `true`) in `struct` view
 - Fixed an exception in `source` view when `prefix` or `postfix` is neither `undefined` nor a string